Add workaround for configuration_aliases tf bug
As noted in the change, there is a bug affecting terraform validate
when used with a module that also requires multiple aliased providers.
There is little activity on this bug, leave it an open question on if
or when it will be fixed. This documents a couple of workaround options
in the interim.