fix(terraform_docs) Add missing popd to code that was introduced in v1.95.0
A bug in the terraform-docs hook was introduced by the code changes in #717.  [A continue statement was added that is not preceded by a `popd` command](https://github.com/antonbabenko/pre-commit-terraform/blob/869a106a4c8c48f34f58318a830436142e31e10a/hooks/terraform_docs.sh#L216).  The bug is only triggered when the hook is processing multiple files where one (or more) of the files does not contain the terraform-docs marker statements.  The hook fails to process the remaining files because the working directory is not reset to the root of the repository.