ci: Disable checks unrunable @ pre-commit.ci (#734)
Enabling faster feedback loop for new contributors by enabling pre-commit.ci in this repo.

The only visible effect of this is when said app is installed and the service is running the checks, it would skip the checks that rely on the external executables that might be missing in runtime. It's necessary since that testing environment is controlled by someone else.