chore(refactoring): Introduce a Python CLI app layout (#738)
* Introduce a Python CLI app layout

This includes a structure with purpose-based modules and a standard
mechanism for adding more subcommands.

When adding a new subcommand, one has to wire the `invoke_cli_app()`
and `populate_argument_parser()` from their new module into the
mappings defined in `_cli_subcommands.py` and `_cli_parsing.py`
respectively. This is the only integration point necessary.

`populate_argument_parser()` accepts a subparser instance of
`argparse.ArgumentParser()` that a new subcommand would need to attach
new arguments into. It does not need to return anything. And the
`invoke_cli_app()` hook is called with an instance of
`argparse.Namespace()` with all the arguments parsed and pre-processed.
This function is supposed to have the main check logic and return an
instance of `._structs.ReturnCode()` or `int`.

* Add a maintainer's manual into the importable package dir

* Simplify subcommand integration to one place

Previously, adding a new subcommand required importing it into two
separate Python modules, adding them into two mappings, maintaining
the same dictionary key string. This is prone to human error and is
underintegrated.

This patch makes it easier by defining a required subcommand module
shape on the typing level. Now, one just need to implement two hook-
functions and a constant with specific signatures and import it in
a single place.

* Drop the `__main__` check per Python docs

See https://docs.python.org/3/library/__main__.html#id1.