 Access `.pre-commit-hooks.yaml` as artifact
Previously, the path lookup logic was unstable due to a flawed
heuristic which meant guessing the artifact path incorrectly in some
cases.

This patch improves that by employing a reliable approach of putting
the artifact into a predictable location within the installed import
package. It then uses a standard `importlib.resources` module to
retrive it from the always-existing place.

The only possibly unobvious quirk is that during development, whenever
this file changes in the Git repository, the `pip install` command
must be executed again since this is what refreshes the file contents
in the installed location.