🧪 Make stderr interceptable
Previously, `sys.stderr` was exposed as a module-scoped var, while
`capsys` patches the `stderr` attribute on the `sys` module. So such
patching did not influence the `stderr` variable in the `_cli` module
as it remained linked to the non-patched object.

With this change, the ability to inspect printing to stderr is
recovered.