🧪 Add initial Python-oriented CI infrastructure
This include jobs building the dists and testing out of them using
pytest in a matrix of supported runtimes. Separate jobs also run
linting.

An additional workflow invokes cron runs. It is separate because GH
is known to disable workflows occasionally. This way, if it disables
the cron workflow, it won't affect the ones running on PRs and pushes.